Moved code from InvoicedemoApplication to openxava
