+++ openxava/pom.xml
@@ -297,6 +297,12 @@
 			<version>4.1.0</version>
 			<scope>provided</scope>
 		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-websocket</artifactId>
+			<version>4.1.0</version>
+			<scope>provided</scope>
+		</dependency>
 
 	</dependencies>
